refactor: split k-point utilities into core package
